+++ CHANGELOG.md
@@ -8,6 +8,7 @@ CHANGELOG
 * instanciate all plugins first, then run their init() functions
 * Important: Please run scripts/migrate/1.8.0.sql and composer install after updating.
 * Chart editor design update
+* default theme for organizations
 
 ### 1.7.11
 * fixed symlink creation in plugin installation